requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
aiofiles==23.2.1
psycopg[binary,pool]>=3.1.0
httpx==0.28.1
sse-starlette==2.2.1
pydantic==2.10.4
databricks-sdk>=0.81.0
